refactor: finalize service and controller after PR review

+++ src/main/java/com/ssafy/sandbox/repository/TodoRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TodoRepository extends JpaRepository<Todo, Integer> {
-    // 추가적인 쿼리가 필요하면 메서드 정의 가능
+public interface TodoRepository extends JpaRepository<Todo, Long> {
 }
